\name{chimeralclustering-package}
\alias{chimeralclustering-package}
\alias{chimeralclustering}
\docType{package}
\title{
\packageTitle{chimeralclustering}
}
\description{
\packageDescription{chimeralclustering}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{chimeralclustering}
\packageIndices{chimeralclustering}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{chimeralclustering}

Maintainer: \packageMaintainer{chimeralclustering}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
